pyproject.toml
[project]
name = "ISPyPSA"
version = "0.1.3beta1"
description = "An open-source capacity expansion model based on the methodology and datasets used by the Australian Energy Market Operator (AEMO) in their Integrated System Plan (ISP)."
authors = [
    { name = "prakaa", email = "abiprakash007@gmail.com" },
    { name = "nick-gorman", email = "n.gorman305@gmail.com" },
    { name = "dylanjmcconnell", email = "dylan.mcconnell@unsw.edu.au" }
]
dependencies = [
    "pandas>=2.2.2,<3.0.0",
    "pypsa==0.34.1",
    "pyyaml>=6.0.2",
    "doit>=0.36.0",
    "xmltodict>=0.13.0",
    "thefuzz>=0.22.1",
    "pyarrow>=18.0.0",
    "isp-trace-parser>=2.0.3",
    "isp-workbook-parser>=2.6.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "pygal>=3.0.5",
    "plotly>=6.5.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.18",
    "numpy>=2.0.0",
    "streamlit>=1.30.0",
    "kaleido>=0.2.1",
]
readme = "README.md"
requires-python = ">=3.11,<3.14"

[project.scripts]
ispypsa = "ispypsa.cli.cli:main"

[project.optional-dependencies]
solvers = [
    "linopy>=0.4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "requests-mock>=1.12.1",
    "ipython>=8.0",
    "ipdb>=0.13.13",
    "pre-commit>=3.8.0",
    "sphinx>=6.0",
    "myst-parser>=3.0.1",
    "myst-nb>=1.1.1",
    "sphinx-copybutton>=0.5.2",
    "sphinx-book-theme>=1.1.3",
    "sphinx-togglebutton>=0.3.2",
    "sphinx-autobuild>=2024.9.3",
    "mkdocs-material>=9.6.9",
    "mkdocs>=1.6.1",
    "mkdocstrings[python]>=0.29.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
# path to tests for pytest
testpaths = ["src", "tests"]
# addopts = add options
# --cov points pytest-cov to the src/ dir
# --cov-branch runs branch coverage
addopts = "-ra --cov=src/ --cov-branch --cov-report xml:tests/coverage.xml --cov-report html:tests/htmlcov"

[tool.coverage.run]
omit = [
    "src/ispypsa/cli/*",
]

[tool.ruff.lint]
select = [
    # isort
    "I",
]

[tool.hatch.build]
include = [
    "src/ispypsa/**/*.py",
    "src/ispypsa/data_fetch/manifests/**/*.txt",
    "src/ispypsa/templater/manually_extracted_template_tables/**/*.csv",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ispypsa"]
exclude = [
    "tests/",
    "docs/",
    "**/__pycache__",
    "*.pyc",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "tests/",
    "docs/",
    ".github/",
    "**/__pycache__",
    "*.pyc",
]
